made balance a double

+++ src/main/java/org/xpdojo/bank/Account.java
@@ -2,9 +2,9 @@ package org.xpdojo.bank;
 
 public class Account {
 
-    private int balance = 0;
+    private double balance = 0;
 
-    public int balance() {
+    public double balance() {
         return balance;
     }
